Major code refactoring
miniplant_reactor.py (needs a better name) now hosts the factory for scenes object,
taking the tilt angle and solar position as parameters.
That's because those are the parameter we need to screen later on.
Basic reactor simulation will similarly have code that takes a location and date/time
and initialize a simulation with the given number of photons.
Finally, the batch_simulation_runner (or something like that) will do the magic of
connecting the dots. Maybe even in a multi-threaded fashion,

Since the final simulations are likely to take long time, it makes sense to have a
raw_data output at the basic_reactor_simulation.py level. Sqlite is an overkill here
so a csv file would be fine (note to self, remember to use safe fai