Replace delegate functions with callable objects
In this way we can go back to use multiprocessing instead of pathos,
i.e. the original pvtrace codebase in the still unreleased feature/cli branch.

This change is important to ensure the code will be eventually compatible with
vanilla upstream pvtrace with no need for local patches.

Incidentally, this also marginally improves performance as multiprocessing is
faster than pathos (see discussion in pvtrace issue #48) IF the env variables
to force numpy to run as single thread are set.