Simplify solar data calculation
This only uses pvlib and removes the solcore dependency.
Algo for solar spectrum is still Spctral2 (also included in pvlib).
Both diffuse and direct spectra are calculated.

TODO: So far we have been adjusting irradiance intensities via ghi/dni/dhi, a more correct approach though would be to re-compoute those in the UV-VIS range that is considered in the simulations... Yes, this is not strictly necessary as the current approximation is likely very similar